ipc: add an exception handler for child processes
This patch introduces a handle_exceptions method, which is called when
a child thread raises an exception and can be overwritten to change the
behavior.

An example user for this would be CLI scripts, wich can print the
exceptions nicely formated.

This approach is similar to what is done in warnings.showwarning[1].

[1] https://docs.python.org/3/library/warnings.html#warnings.showwarning

Signed-off-by: Filipe Laíns <lains@archlinux.org>