ipc: add vital argument to ProcessManager.start
This patch introduces a 'vital' keyword argument to
ProcessManager.start, which will exit the current process if
the child processes exit with a bad (non-zero) code.

Chances are if there is an issue in the child process, something went
wrong, and we should abort. There are, however, valid cases where this
might happen, and probably you have a custom way of handling them. For
this reason, the argument is set by default, if you have one of those
special cases where you want to handle this on your own, the behavior
can naturally be disabled.

I am not a super fan of this implementation, which spins up a thread to
watch the process states and exit the main process in case something
goes wrong. However, this is by far the best way I could get it to work.
The thread loop runs on a delay, so it won't eat up your precious cpu
time :P

Signed-off-by: Filipe Laíns <lains@archlinux.org>