analyzer/ipc: add a way to send exceptions between IPC processes
This patch introductes a way to send exceptions between processes. It
uses pipes to acheive it. Each process manager can receive a input and
output connection to exchange exceptions.

The goal here is to allow the frontend to manage the backend exceptions.
We pass the ouputput exception pipe from the backend to the frontend.
This allows the frontend to receive exceptions from the backend and show
them.

Signed-off-by: Filipe Laíns <lains@archlinux.org>