setup.py: remove unecessary data instalation
setuptools' data_files option is meant to install files to the package
data directory, not to make sure they are included, that should happen
by default if the module they are in is being installed.

Files installed this way can be accessed with pkg_resources, or derived
from importlib.metadata. But we do not do this, we access the file as
relative to __file__.

Furthermore, data_files seems to be deprecated. The official
documentation says it does not works with wheels but it looks like it
does here ¯\_(ツ)_/¯

Signed-off-by: Filipe Laíns <lains@archlinux.org>