ipc/analyzer: let backends signal when they are ready
Currently, frontends such as the Qt one will be shown before the backend
is ready and we won't have any information about what is hapenning. We
just have to wait until data shows up, and since we have backends such
as Luna, that take a while to get setup, this is not a great UX.

This patch adds two things, 1) a way for the backend to signal when it
is ready, and 2) a way for backends to send setup messages to the
frontend, so that we can tell the users what is hapenning (eg. building
gateware, waiting for device to show up, etc.)

It also adds a few callbacks in ViewSBFrontend to interact with the
added interfaces. The default implementation will print setup messages
to stdout.

Signed-off-by: Filipe Laíns <lains@riseup.net>